+++ docs/CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+* In dashboard, if a category is selected and the back button is pressed,
+the category is unselected
 
+
+### Changed
+* Don't show "No category" in dashboard if it's empty
 
 
 ## 1.9.2 - 2019-10-29